requirements.txt
pyserial
flask
time
scipy.interpolate
matplotlib.pyplot
numpy